write out session attribute to the dynamo_db (#24)
I think we should all talk about this, but I was thinking that we should:
*prompt the user for some defaults the first time they login. Yeah it's not ideal, but it will be easier for us to do than to allow them to say "no I don't want to set defaults" So I say we just make them set a least some subset of defaults :).

Can you explain your reserved issues more? 

I understand "how to use the information saved in the table" but what do you mean by "bring up suggestion directly without asking user for information"?

Regarding fields in the dataset, I actually thought about it a lot. Having a NoSQL database is kind of restrictive - they have to be flat, you can't have lists, but I designed this one for the ways we will query it. So I do not think this is a work item. 
It's fine if users visit several restaurants, there should just be multiple entries, since the sort key is different. That is why it is unique.
See example table:
https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/LSI.html


Regarding: how to apply database knowledge to do some recommendation with other user's information(like two user visited the same restaurant, can we recommend them each other's history?)

I think this is captured in the Trends table as aggregate information:
https://docs.google.com/document/d/1oRrgc1A0x9ZgDc8cvGhX_0_sIH5e4vaFHg8f_RZnTDY/edit#heading=h.79cdq4lrsofs